Add update checker
Add an update checker which grabs the latest version from GitHub and checks if it's newer than the running process.
It's designed so that even in a critical failure, such as when the game updates and SEWT can't even initialize, it can still check for an update and notify the user.