Enable deterministic builds.
Added SourceLink support, and moved version information to Directory.Build.props. This allows the version to be overridden at build-time. Informational version (shown in explorer) now also includes the git commit hash.
This means a build with the exact same sources will always produce byte-level perfectly identical outputs.
This is an addition to the alternative implementation for #34.