Grab DS from Steam if game not found during build
If a Steam install of SE can't be found, LinkBinaries.ps1 will try to download the DS binaries from Steam. Steam client is not required to do this. This only downloads the .DLL files, so it's reasonably fast.

For build performance, the DS binaries will not auto-update; execute this script with --Force to update the DS manually.

Synced .gitignore with root, and removed the one from project.

Make sure msbuild call restores nuget packages. Initialize the entire dev environment, if necessary, instead of just calling msbuild directly so the build-time environment is stable.