Fix description not applying with change verb.
Add a warning if the description exceeds Steam's limit of 8000 bytes.
Since Steamworks.NET sends the data over as UTF-8, this will test the UTF-8 size, to accommodate various encodings and character content.

For example, a 16000 byte UTF-16 encoded file, with only 7-bit ASCII characters, will be 8000 bytes once converted to UTF-8, and is valid.