Check for empty or missing Data folder, fixes #37
If the Data directory is missing, or is empty, it won't be published, so warn for that.
If --force was specified, create a placeholder file (and create the directory if necessary) to ensure it'll be uploaded.
SE will throw an exception if the Data directory isn't present.

Adjust logging to always prepend ERROR or WARNING when using calling the respective logging methods.